Split native Zenmap profile argument helpers
